betty issues

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -10,7 +10,7 @@ hash_table_t *hash_table_create(unsigned long int size)
 {
 	/* creates a new hash table */
 	hash_table_t *newTable;
-	unsigned int i = 0;
+	unsigned long int i = 0;
 
 	/* allocate memory for hash table */
 	newTable = malloc(sizeof(hash_table_t));
